backend/automation/requirements.txt
torch
torchvision
torchaudio
opencv-python
pillow
numpy
matplotlib
scikit-learn
tqdm
ftfy
regex
git+https://github.com/openai/CLIP.git
faiss-cpu
